Sign of the rotations
The sign of the rotations in langevin.py and gmx rotmat
(and rotdiff.py) are opposite.
Langevin.py : current angle formed with the x axis
gmx rotmat : the angle needed for superposition onto
            the reference

In langevin.py, there needs to be 3 particles in order
to correctly calculate the rotation matrix with gmx rotmat.